requirements.txt
discord.py
python-dotenv
requests
pandas
gspread
google-auth
openai
unbelievaboat
asyncio
json
shlex
random